+++ src/renderers/dom/client/eventPlugins/ChangeEventPlugin.js
@@ -1,10 +1,8 @@
 /**
- * Copyright 2013-present, Facebook, Inc.
- * All rights reserved.
+ * Copyright (c) 2013-present, Facebook, Inc.
  *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree. An additional grant
- * of patent rights can be found in the PATENTS file in the same directory.
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
  *
  * @providesModule ChangeEventPlugin
  */
@@ -156,7 +154,7 @@ if (ExecutionEnvironment.canUseDOM) {
 
   isInputEventSupported =
     isEventSupported('input') &&
-    (!('documentMode' in document) || document.documentMode > 9);
+    (!document.documentMode || document.documentMode > 9);
 }
 
 /**